mv: decode and normalize move to StringHandler, normalize() -> normalize_to_absolute_path()
